This file contains five versions of UDF to check if a given number is prime or not.
The first three versions are copies of a YouTube Python teaching video whose link is inside the file comment.

The fourth version is created by me based off of the third version, replacing the for loop with a while loop.

The fifth version is derived from:
https://www.sanfoundry.com/python-program-compute-prime-factors-integer/

The fifth version is really incredible since it goes a totally different way compared to the previous four versions.