Playing_with_digits from codewars.com
This is a program to find out an integer k (if it exists) which a given integer n is multiplied by to get the sum of
the given integer's own digits each raised to the power which comes from a series of numbers starting from p and
ending with p + (d - 1). 'd' stands for the number of the given integer's digits.