This files contains a program which helps us solve Exercise 2.1-19 in Thomas' Calculus.
The rule for computing the values of h or Δx has been rewritten, which
looks more like the general rule for a geometric sequence.

Notes: The general rule for a geometric sequence is Xn = ar^(n-1) where
'a' is the first term, 'r' is the common ratio, and 'n' is the ordinal
number.